Cleansup TabScreen to separate components
